requirements.txt
streamlit==1.38.0
python-docx==1.1.2
streamlit-option-menu==0.3.13
reportlab==4.2.2
PyMuPDF==1.24.10  # provides 'fitz' module for PDF text extraction
tiktoken
google-generativeai==0.8.2
python-dotenv
pytest
st-copy-to-clipboard==0.1.6
requests>=2.31.0
pyarrow==17.0.0

